requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
autogen-agentchat>=0.4.0
autogen-core>=0.4.0
autogen-ext>=0.4.0
openai>=1.0.0
matplotlib==3.7.3
pandas==2.0.3
networkx==3.1
requests==2.31.0 